Rename typescript helper file to follow convention
